Move loading spinner to components directory
